requirements.txt
# Core Dependencies
pandas>=1.3.0
numpy==1.26.0
ccxt>=2.0.0
requests>=2.26.0

# Machine Learning
scikit-learn>=1.0.1
ta-lib>=0.4.20

# Optional: Sentiment Analysis
tweepy>=4.0.0
praw>=7.5.0
newsapi-python>=0.2.6
tensorflow-intel==2.15.0